refactor(tui): delete local mode; the TUI is always an API client
Plan 018 C4. Model, its Update/View/Init, the local subscription
forwarders, tickCmd/TickMsg, and TUILocalTickInterval are gone; the tui
package no longer touches the supervisor or log manager directly.
Survivors relocated: shared msg types and detail-view data types to
client.go, Mode/ViewMode + nearBottomThreshold + the process style
helpers to base.go. convertRequestRecordToDetail and the capture-dir
helpers died with their only caller (local Enter/detail path);
base_convert_test.go went with them. TestClientModel_NeverPollsProcesses
now uses a local unhandled msg type instead of TickMsg.

Note: internal/tui still pulls internal/supervisor and internal/logs
TRANSITIVELY via internal/api's wire types — pre-existing (predates this
plan), verified unrelated to local mode; direct imports are zero.

Net-deletion commit: simplify and external review skipped per flow
profile; gate green under -race.
