+++ internal/constants/constants.go
@@ -194,17 +194,6 @@ const (
 	StreamReconnectFlapThreshold = 1 * time.Second
 )
 
-// TUI timing
-const (
-	// TUILocalTickInterval drives the local-mode TUI's periodic process-list
-	// refresh, and only local mode's: it reads the in-process supervisor
-	// directly, so a fast tick costs nothing. Attach mode used to share the
-	// interval for an HTTP poll; C12 deleted that poll in favor of the
-	// processes SSE stream, so nothing in attach mode ticks any more
-	// (plan 017 C12).
-	TUILocalTickInterval = 500 * time.Millisecond
-)
-
 // Log configuration
 const (
 	// DefaultLogLimit is the default number of log lines to return
