--- internal/tui/base_convert_test.go
@@ -1,150 +1,0 @@
-package tui
-
-import (
-	"bytes"
-	"compress/gzip"
-	"os"
-	"path/filepath"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/charliek/prox/internal/proxy"
-)
-
-func gzipBytes(t *testing.T, data []byte) []byte {
-	t.Helper()
-	var buf bytes.Buffer
-	gw := gzip.NewWriter(&buf)
-	_, err := gw.Write(data)
-	require.NoError(t, err)
-	require.NoError(t, gw.Close())
-	return buf.Bytes()
-}
-
-// TestConvertRequestRecordToDetail_FilePathBacked pins that a FilePath-backed
-// (disk-spilled) body is loaded — previously the TUI conversion read Data
-// directly and dropped disk bodies.
-func TestConvertRequestRecordToDetail_FilePathBacked(t *testing.T) {
-	dir := t.TempDir()
-	payload := []byte(`{"disk":"backed"}`)
-	fp := filepath.Join(dir, "res.bin")
-	require.NoError(t, os.WriteFile(fp, payload, 0600))
-
-	rec := proxy.RequestRecord{
-		ID:        "disk001",
-		Timestamp: time.Now(),
-		Method:    "GET",
-		URL:       "/data",
-		Details: &proxy.RequestDetails{
-			ResponseBody: &proxy.CapturedBody{
-				Size:         int64(len(payload)),
-				CapturedSize: int64(len(payload)),
-				ContentType:  "application/json",
-				FilePath:     fp,
-			},
-		},
-	}
-
-	detail := convertRequestRecordToDetailWithDirs(rec, []string{dir})
-	require.NotNil(t, detail.ResponseBody)
-	assert.False(t, detail.ResponseBody.Unavailable)
-	assert.False(t, detail.ResponseBody.IsBinary)
-	assert.Equal(t, string(payload), detail.ResponseBody.Data)
-}
-
-// TestConvertRequestRecordToDetail_Stale verifies local-mode conversion
-// (Model, which holds real proxy.RequestRecord values) computes Stale itself
-// via RequestRecord.StaleAt rather than needing it threaded in separately
-// (D8, #53), and that a fresh in-flight record is not marked stale.
-func TestConvertRequestRecordToDetail_Stale(t *testing.T) {
-	staleRec := proxy.RequestRecord{
-		ID:        "stale1",
-		Timestamp: time.Now().Add(-10 * time.Minute),
-		Method:    "GET",
-		URL:       "/stream",
-		InFlight:  true,
-	}
-	detail := convertRequestRecordToDetail(staleRec)
-	assert.True(t, detail.Stale)
-
-	freshRec := proxy.RequestRecord{
-		ID:        "fresh1",
-		Timestamp: time.Now(),
-		Method:    "GET",
-		URL:       "/stream",
-		InFlight:  true,
-	}
-	detail = convertRequestRecordToDetail(freshRec)
-	assert.False(t, detail.Stale)
-}
-
-// TestConvertCapturedBodyToBodyData_InvalidUTF8NotStringified pins the TUI's
-// defense-in-depth: bytes that are not valid UTF-8 are never treated as text
-// for rendering, even when the stored record claims IsBinary=false (e.g. a
-// socket-supplied flag or a disk file mutated after classification) — they
-// are reclassified binary instead. The raw bytes are still carried in
-// bd.Data byte-for-byte (a Go string is just bytes) so the hexdump preview
-// (D11, #50.2) has something to dump; what must never happen is those bytes
-// reaching the plain-text render path.
-func TestConvertCapturedBodyToBodyData_InvalidUTF8NotStringified(t *testing.T) {
-	raw := []byte{'h', 'i', 0xFF, 0xFE}
-	body := &proxy.CapturedBody{
-		Size:         4,
-		CapturedSize: 4,
-		ContentType:  "text/plain",
-		IsBinary:     false, // lies: data is not valid UTF-8
-		Data:         raw,
-	}
-
-	bd := convertCapturedBodyToBodyData(body, nil)
-	require.NotNil(t, bd)
-	assert.True(t, bd.IsBinary, "invalid UTF-8 must be reclassified binary")
-	assert.Equal(t, string(raw), bd.Data, "raw bytes are preserved for the hexdump preview")
-}
-
-func TestConvertRequestRecordToDetail_GzipFilePath(t *testing.T) {
-	dir := t.TempDir()
-	payload := []byte(`{"gzip":"on disk"}`)
-	fp := filepath.Join(dir, "res.bin")
-	require.NoError(t, os.WriteFile(fp, gzipBytes(t, payload), 0600))
-
-	rec := proxy.RequestRecord{
-		ID: "gz1",
-		Details: &proxy.RequestDetails{
-			ResponseBody: &proxy.CapturedBody{
-				ContentType:     "application/json",
-				ContentEncoding: "gzip",
-				IsBinary:        true,
-				FilePath:        fp,
-			},
-		},
-	}
-
-	detail := convertRequestRecordToDetailWithDirs(rec, []string{dir})
-	require.NotNil(t, detail.ResponseBody)
-	assert.False(t, detail.ResponseBody.IsBinary)
-	assert.Equal(t, string(payload), detail.ResponseBody.Data)
-}
-
-func TestConvertRequestRecordToDetail_EvictedFilePath(t *testing.T) {
-	dir := t.TempDir()
-	rec := proxy.RequestRecord{
-		ID: "ev1",
-		Details: &proxy.RequestDetails{
-			ResponseBody: &proxy.CapturedBody{
-				Size:        10,
-				ContentType: "application/json",
-				FilePath:    filepath.Join(dir, "gone.bin"),
-			},
-		},
-	}
-
-	detail := convertRequestRecordToDetailWithDirs(rec, []string{dir})
-	require.NotNil(t, detail.ResponseBody)
-	assert.True(t, detail.ResponseBody.Unavailable)
-	assert.Equal(t, "evicted", detail.ResponseBody.UnavailableReason)
-	assert.Empty(t, detail.ResponseBody.Data)
-}
