+++ internal/tui/client.go
@@ -2,6 +2,7 @@ package tui
 
 import (
 	"errors"
+	"time"
 
 	tea "github.com/charmbracelet/bubbletea"
 
@@ -10,6 +11,97 @@ import (
 	"github.com/charliek/prox/internal/proxy"
 	"github.com/charliek/prox/internal/stream"
 )
+
+// LogEntryMsg is sent when a new log entry arrives
+type LogEntryMsg domain.LogEntry
+
+// ProxyRequestMsg is sent when a new proxy request is recorded
+type ProxyRequestMsg proxy.RequestRecord
+
+// ProcessesMsg is sent when processes should be refreshed
+type ProcessesMsg []domain.ProcessInfo
+
+// RestartResultMsg is sent when a restart operation completes
+type RestartResultMsg struct {
+	Process string
+	Err     error
+}
+
+// RestartResultClearMsg is sent to clear the restart result after a delay
+type RestartResultClearMsg struct{}
+
+// RequestDetailMsg is sent when request details are loaded. Seq is the
+// fetchRequestDetail call's sequence number (ClientModel.detailFetchSeq at
+// dispatch time) — the handler drops any msg whose Seq doesn't match the
+// current value, so a stale or superseded fetch can never clobber a newer
+// one (D16).
+type RequestDetailMsg struct {
+	ID      string
+	Seq     int
+	Details *RequestDetailData
+}
+
+// RequestDetailErrorMsg is sent when loading request details fails. Seq is
+// the fetchRequestDetail call's sequence number (see RequestDetailMsg).
+type RequestDetailErrorMsg struct {
+	ID  string
+	Seq int
+	Err error
+}
+
+// RequestDetailData holds the detailed information about a request for TUI display
+type RequestDetailData struct {
+	ID         string
+	Timestamp  string
+	Method     string
+	URL        string
+	Subdomain  string
+	StatusCode int
+	DurationMs int64
+	RemoteAddr string
+	// InFlight marks a request whose response is still streaming: Duration
+	// renders as "(in flight)" and a nil Details gets an explanatory note
+	// instead of being treated as "capture not enabled".
+	InFlight bool
+	// Stale marks an in-flight request that has been in-flight longer than
+	// constants.InFlightStaleAfter (D8, #53): the completion event may have
+	// been lost and the true outcome is unknown. Always false when InFlight
+	// is false.
+	Stale           bool
+	RequestHeaders  map[string][]string
+	ResponseHeaders map[string][]string
+	RequestBody     *BodyData
+	ResponseBody    *BodyData
+}
+
+// BodyData holds captured body information
+type BodyData struct {
+	Size            int64
+	Truncated       bool
+	ContentType     string
+	ContentEncoding string
+	IsBinary        bool
+	Data            string
+	// DataBase64 marks Data as base64-encoded (the wire format: the API
+	// base64-encodes binary bodies). A producer that already holds raw bytes
+	// leaves this false, so the hex preview never has to guess — raw bytes
+	// that merely LOOK like base64 must not be decoded.
+	DataBase64 bool
+	// Unavailable is true when the body existed but could no longer be loaded
+	// (e.g. its disk file was evicted); UnavailableReason explains why.
+	Unavailable       bool
+	UnavailableReason string
+}
+
+// restartResultClearDelay is how long to show restart result before clearing
+const restartResultClearDelay = 3 * time.Second
+
+// restartResultClearCmd returns a command that clears the restart result after a delay
+func restartResultClearCmd() tea.Cmd {
+	return tea.Tick(restartResultClearDelay, func(t time.Time) tea.Msg {
+		return RestartResultClearMsg{}
+	})
+}
 
 // ClientOptions carries everything that differs between the two client-mode
 // callers — `prox attach`, talking to another process's daemon over HTTP, and
@@ -179,8 +271,7 @@ func (m ClientModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		m.handleLogEntry(domain.LogEntry(msg))
 
 	case LogsSyncMsg:
-		// One batch, one render (C9). Attach mode only: local mode reads the
-		// log manager's ring directly and has nothing to synchronize against.
+		// One batch, one render (C9).
 		m.handleLogsSync(msg)
 
 	case ProxyRequestMsg:
@@ -197,8 +288,7 @@ func (m ClientModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		}
 
 	case RequestsSyncMsg:
-		// One batch, one render (C6). Attach mode only: local mode reads the
-		// request ring directly and has nothing to synchronize against.
+		// One batch, one render (C6).
 		m.handleRequestsSync(msg)
 		// D16 also applies to a completion that arrives via the sync batch
 		// instead of a live event: an open detail still showing an in-flight
